feat: update linter rules
