--- lib/constants/constants.dart
@@ -1,5 +1,0 @@
-// Copyright 2022 The AnimatedSvg Author. All rights reserved.
-// Use of this source code is governed by a MIT-style license that
-// can be found in the LICENSE file.
-
-export 'string.dart';
