--- lib/constants/string.dart
@@ -1,9 +1,0 @@
-// Copyright 2022 The AnimatedSvg Author. All rights reserved.
-// Use of this source code is governed by a MIT-style license that
-// can be found in the LICENSE file.
-
-// Strings
-
-/// Assertion message to show required children count.
-const kRequiredChildrenAssertMessage =
-    'AnimatedSvg: It is required to provide only 2 SvgPicture as children!';
