+++ lib/src/animated_svg_controller.dart
